+++ performance/os-performance/go.mod
@@ -3,6 +3,7 @@ module github.com/OpenSlides/OpenSlides/performance/os-performace
 go 1.15
 
 require (
+	github.com/spf13/cobra v1.1.3
 	github.com/vbauerster/mpb/v6 v6.0.2
 	golang.org/x/sys v0.0.0-20210218155724-8ebf48af031b // indirect
 )
